refactoring xdg functions
