+++ Makefile
@@ -3,13 +3,15 @@ SRCDIR :=	src/
 INCLUDE :=	include/
 
 CXX_SRC :=	main.cpp \
-		Server.cpp
+		Server.cpp \
+		XdgShell.cpp
 
 C_SRC :=	xdg-shell-protocol.c
 
 CXX :=		g++
 CXXFLAGS :=	-W -Wall -Wextra -g -std=c++17 \
 		-lwayland-server -lwlroots -DWLR_USE_UNSTABLE
+
 CXX_SRC :=	$(addprefix $(SRCDIR), $(CXX_SRC))
 C_SRC :=	$(addprefix $(SRCDIR), $(C_SRC))
 CXX_OBJ :=	$(CXX_SRC:.cpp=.o)
